+++ backtester/return_calc/log_return.py
@@ -1,5 +1,6 @@
 import quantkit.backtester.return_calc.return_metrics as return_metrics
 import quantkit.mathstats.mean.rolling_mean as rolling_mean
+import quantkit.mathstats.mean.numpy_mean as numpy_mean
 import quantkit.utils.annualize_adjustments as annualize_adjustments
 import numpy as np
 import pandas as pd
@@ -23,7 +24,7 @@ class LogReturn(return_metrics.ReturnMetrics):
     def __init__(self, universe: list, frequency: str = None, **kwargs) -> None:
         super().__init__(universe)
         self.frequency = frequency
-        self.return_calculator = rolling_mean.RollingMean(
+        self.return_calculator = numpy_mean.NumpyWindowMean(
             num_ind_variables=self.universe_size, **kwargs
         )
 
@@ -102,7 +103,9 @@ class LogReturn(return_metrics.ReturnMetrics):
 
         outgoing_row = np.squeeze(self.return_calculator.windowed_outgoing_row)
         self.return_calculator.update(
-            np.log(annualized_return + 1), outgoing_row, index=date
+            incoming_variables=np.log(annualized_return + 1),
+            outgoing_variables=outgoing_row,
+            index=date,
         )
 
     def is_valid(self):